+++ NEWS.md
@@ -11,6 +11,8 @@
 
 * Fixed bug in print outputs. Row numbers are now kept as their removal was causing the issue. This is a temporary fix; see [#60](https://github.com/mikemc/speedyseq/issues/60).
 
+* Add  `as_tibble()` methods for phyloseq objects
+
 # speedyseq 0.5.3
 
 * `print(physeq)` now aligns the `refseq(physeq)` summary properly